data structure options : verb conjugation
Trying out two possible configurations for the template for verb conjugation.
The first - <verb conjugation_A.csv> - would contain one record/row for each verb in a Dialect, with all the verb tenses represented as columns. There are separate adjacent columns for the English translation.
The second option - <verb conjugation_B.csv> - uses fewer columns by storing the tense as an option in a single field (column) — but more rows, by having one record for each distinct Dialect-conjugation possibility.

As of right now I think that Option B will have more flexibility down the line, because it seems like it should also allow for possible alternate translations or uses
Option B also seems to have a smaller file size, and seems to be more readable to the human eye.
However, what remains to be seen is whether the performance efficiency is optimal in machine-readable form, because a database may need to perform lookups based on the corresponding verb tense.